requirements.txt
asgiref==3.4.1
astroid==2.8.4
autopep8==1.6.0
cachetools==4.2.4
certifi==2018.8.24
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.5.0
Django==2.2.13
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gunicorn==19.7.1
httplib2==0.20.2
idna==2.7
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==2.1.0
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.11.1
pyparsing==3.0.6
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.7.2
six==1.11.0
sqlparse==0.3.0
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.24.2
whitenoise==3.3.1
wrapt==1.13.3
oauth2client
httplib2